Separate recovery word count example
