+++ examples/wordcount.py
@@ -1,24 +1,11 @@
 import re
 
-from bytewax import AdvanceTo, Dataflow, Emit, parse, spawn_cluster
-from bytewax.recovery import KafkaRecoveryConfig
+from bytewax import Dataflow, parse, run_cluster
 
 
-def input_builder(worker_index, worker_count, resume_epoch):
-    with open("examples/sample_data/wordcount.txt") as lines:
-        for epoch, line in enumerate(lines):
-            if epoch < resume_epoch:
-                continue
-            if epoch % worker_count != worker_index:
-                continue
-            # if epoch == 12:
-            #     raise RuntimeError("boom")
-            yield AdvanceTo(epoch)
-            yield Emit(line)
-
-
-def output_builder(worker_index, worker_count):
-    return print
+def file_input():
+    for line in open("examples/sample_data/wordcount.txt"):
+        yield 1, line
 
 
 def lower(line):
@@ -33,13 +20,8 @@ def initial_count(word):
     return word, 1
 
 
-def count_builder(word):
-    return 0
-
-
-def add(running_count, new_count):
-    running_count += new_count
-    return running_count, running_count
+def add(count1, count2):
+    return count1 + count2
 
 
 flow = Dataflow()
@@ -50,19 +32,11 @@ flow.flat_map(tokenize)
 # "words"
 flow.map(initial_count)
 # ("word", 1)
-flow.stateful_map("running_count", count_builder, add)
-# ("word", running_count)
+flow.reduce_epoch(add)
+# ("word", count)
 flow.capture()
 
 
 if __name__ == "__main__":
-    recovery_config = KafkaRecoveryConfig(
-        ["localhost:9092"], "bytewax-state", create=True
-    )
-    spawn_cluster(
-        flow,
-        input_builder,
-        output_builder,
-        recovery_config=recovery_config,
-        **parse.cluster_args()
-    )
+    for epoch, item in run_cluster(flow, file_input(), **parse.cluster_args()):
+        print(epoch, item)
